ona/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-telegram-bot==20.8
httpx==0.26.0
pydantic==2.5.0
python-dotenv==1.0.0
supabase==2.3.0
openai==1.79.0
pytest>=7.3.1 